+++ CHANGELOG.md
@@ -5,6 +5,11 @@ All notable changes to this crate will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## v1.0.0 - 2023-02
+## v1.0.1 - 2023-03-01
+
+- Add const BYTES ProofOfKnowledge structs
+- Add to_bytes and from_bytes to ProofOfKnowledge structs
+
+## v1.0.0 - 2023-02-28
 
 - Initial release.